Program.cs
I chose to use C#.Net because I am most familiar / comfortable with it.  

Normally, I would include SequenceFixer into its own class file but for demonstration purposes I have included the class within the main file, Program.cs

The program will accept a filename as a command-line argument.  If the argument is not found, the program will prompt the user for a path to the input file.

I had a lot of fun with this project.  I believe I've represented my programming style very well.  Thank you again for this opportunity.